backend/python-django/requirements.txt
Django==6.0.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
mysql-connector-python==9.3.0
PyJWT==2.10.1
bcrypt==4.3.0
gunicorn==23.0.0
